+++ src/test/java/com/codedifferently/labs/partA/ex01/SavingsAccountTest.java
@@ -1,5 +1,6 @@
 package com.codedifferently.labs.partA.ex01;
 
+import org.junit.jupiter.api.Test;
 import partA.ex01.SavingAccountExample;
 
 public class SavingsAccountTest {
@@ -8,6 +9,7 @@ public class SavingsAccountTest {
         sa0001.name = "Damien";
         sa0001.interestRate = 0.02;
         sa0001.deposit(1000);
+        System.out.println(sa0001.balance);
 
         SavingAccountExample sa0002 = new SavingAccountExample();
         sa0002.balance = 2000;
