+++ src/test/java/com/codedifferently/labs/partB/CatTest.java
@@ -1,4 +1,7 @@
 package com.codedifferently.labs.partB;
 
 public class CatTest {
+    public void setName(String name){
+
+    }
 }
